chat_prj/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.1
django-appconf==1.0.6
django-channels==0.7.0
django-imagekit==5.0.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
oauthlib==3.2.2
pilkit==3.0
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==7.0.3
